portable snprintf abstraction for non posix platforms

+++ libharvid/ffdecoder.c
@@ -255,9 +255,9 @@ static void ff_init_moviebuffer(void *ptr) {
   ff->buf_height = ff->out_height;
   if (!ff->buffer) {
 #ifdef WIN32
-    fprintf(stderr, "out of memory (trying to allocate %lu bytes)\n", numBytes); // XXX
+    fprintf(stderr, "out of memory (trying to allocate %lu bytes)\n", (long unsigned) numBytes);
 #else
-    fprintf(stderr, "out of memory (trying to allocate %zu bytes)\n", numBytes); // XXX
+    fprintf(stderr, "out of memory (trying to allocate %zu bytes)\n", numBytes);
 #endif
     exit(1);
   }
